mudança de entidades

+++ src/main/java/com/senai/bibliotecaemprestimo/repositories/AdminRepository.java
@@ -8,5 +8,5 @@ import java.util.Optional;
 
 @Repository
 public interface AdminRepository extends JpaRepository<Admin, Long> {
-    Optional<Admin> findById(int id);
+    Optional<Admin> findById(Long id);
 }
